testCacheExpires only for Caches where Yii controls expiring.
