+++ tests/unit/framework/caching/FileCacheTest.php
@@ -2,6 +2,7 @@
 namespace yiiunit\framework\caching;
 use yii\caching\FileCache;
 use yiiunit\TestCase;
+use yii\caching\TimeProvider;
 
 /**
  * Class for testing file cache backend
@@ -22,4 +23,17 @@ class FileCacheTest extends CacheTest
 		}
 		return $this->_cacheInstance;
 	}
+
+    public function testExpire()
+    {
+        $cache = $this->getCacheInstance();
+        $now = time();
+        TimeProvider::setTime($now);
+        $this->assertTrue($cache->set('expire_test', 'expire_test', 2));
+        TimeProvider::setTime($now+1);
+        $this->assertEquals('expire_test', $cache->get('expire_test'));
+        TimeProvider::setTime($now+3);
+        $this->assertEquals(false, $cache->get('expire_test'));
+        TimeProvider::setTime(null);
+    }
 }
